Chargement diapos + images fctionnels, reste un probleme dans la trasmission des infos modele -> vue pour les images

+++ V6_MVP_lecteurDiaporama/Fichiers de code/choixvitessedefilement.h
@@ -23,7 +23,7 @@ private slots:
     void demanderEnvoiValeur();
 
 signals:
-    void envoyerVitesseDfl(float);
+    void envoyerVitesseDfl(unsigned int);
 };
 
 #endif // CHOIXVITESSEDEFILEMENT_H
